+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Added
 
 - `delegate!` macro to reduce code duplication in the `Animation` enum implementation.
+- AnimationKind enum to track the state transitions between animations.
 
 ### Changed